InboxAdminComponent: listing not resolved inbox messages

+++ smsgw/resources/inbox/api.py
@@ -7,6 +7,7 @@ from flask.ext.classy import FlaskView, route
 from smsgw.models import Inbox, User
 from smsgw.lib.utils import response
 from smsgw.resources import decorators
+from smsgw.resources.error.api import ErrorResource
 from smsgw.extensions import db
 
 
@@ -15,6 +16,7 @@ class InboxResource(FlaskView):
 
     route_base ='/'
 
+    @route('/inbox/', methods=['GET'])
     @route('/users/<uuid:user_uuid>/inbox/', methods=['GET'])
     @route('/users/<uuid:user_uuid>/applications/<uuid:application_uuid>/inbox/',
            methods=['GET'])
@@ -25,14 +27,21 @@ class InboxResource(FlaskView):
         """
         user = kwargs.get('user')
         app = kwargs.get('application')
+        query = None
 
-        # make query
-        query = app.inbox if app else user.inbox
-        query = query.order_by(Inbox.received.desc())
+        if user is None:
+            query = Inbox.query.order_by(Inbox.received.desc())
+
+            if not request.user.is_admin():
+                raise ErrorResource(message='Not have permissions.', status_code=403)
+        else:
+            query = app.inbox if app else user.inbox
+            query = query.order_by(Inbox.received.desc())
 
         return response([message.to_dict() for message in query.all()])
 
 
+    @route('/inbox/<uuid:inbox_uuid>/', methods=['DELETE'])
     @route('/users/<uuid:user_uuid>/inbox/<uuid:inbox_uuid>/', methods=['DELETE'])
     @route('/users/<uuid:user_uuid>/applications/<uuid:application_uuid>/inbox/<uuid:inbox_uuid>/',
            methods=['DELETE'])
@@ -41,7 +50,13 @@ class InboxResource(FlaskView):
         """
         Delete user contact
         """
+        user = kwargs.get('user', request.user)
         inbox = kwargs.get('inbox')
+
+        # admin check, to be sure that non admin users can delete inboxes
+        if inbox.userId != user.id and not user.is_admin():
+            raise ErrorResource(message='Not have permissions.', status_code=403)
+
         db.session.delete(inbox)
         db.session.commit()
